Work on consolidating the format utilities into the package format.
