+++ format/format.go
@@ -1,0 +1,3 @@
+// Package format implements functions to format numbers,
+// dates, times and strings.
+package format
